Squash Merge 'master' into 'animated-real-time-feedback'
