+++ ProjectSpace/commsgaze/build.gradle
@@ -24,7 +24,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
 }
+
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
@@ -43,5 +45,4 @@ dependencies {
     annotationProcessor 'androidx.room:room-compiler:2.2.5'
     implementation 'androidx.recyclerview:recyclerview:1.0.0'
     annotationProcessor 'android.arch.persistence.room:compiler:1.1.1'
-
 }
