+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/Direction.java
@@ -7,3 +7,4 @@ package com.pwc.commsgaze;
 public enum Direction {
     TOP,LEFT,NEUTRAL,RIGHT,BOTTOM,UNKNOWN;
 }
+
