+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/MainRecyclerViewAdapter.java
@@ -3,18 +3,22 @@ package com.pwc.commsgaze;
 import android.media.AudioAttributes;
 import android.media.MediaPlayer;
 import android.net.Uri;
+import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
 import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
+
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 import com.pwc.commsgaze.database.Content;
+
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
@@ -22,7 +26,19 @@ import java.util.List;
 public class MainRecyclerViewAdapter extends RecyclerView.Adapter<MainRecyclerViewAdapter.ViewHolder> {
     public final static String TAG = "MainRecyclerViewAdapter";
     private List<Content> contents;
+    private int deviceHeight;
+    private int deviceWidth;
+    private int fixedDimension;
 
+    MainRecyclerViewAdapter(DisplayMetrics displayMetrics,int fixedDimension){
+        /*TODO to consider for vertical orientation as well*/
+        deviceHeight = displayMetrics.heightPixels  ;
+        deviceWidth = displayMetrics.widthPixels;
+        Log.d(TAG,"Device Height "+ deviceHeight);
+        Log.d(TAG,"Device Width "+ deviceWidth);
+        this.fixedDimension = fixedDimension;
+
+    }
 
     @NonNull
     @Override
@@ -36,10 +52,12 @@ public class MainRecyclerViewAdapter extends RecyclerView.Adapter<MainRecyclerVi
     public void onBindViewHolder(@NonNull MainRecyclerViewAdapter.ViewHolder holder, int position) {
         holder.textView.setText(contents.get(position).getWord());
         Uri imageUri = Uri.fromFile(new File(contents.get(position).getImageDirPath()));
+        int suitableDimension = deviceWidth/(fixedDimension+2);
         Glide.with(holder.itemView.getContext())
                 .load(imageUri)
-                .apply(new RequestOptions().override(250, 250))
+                .apply(new RequestOptions().override(suitableDimension,suitableDimension))
                 .into(holder.imageView);
+
         holder.setAudioDirPath(contents.get(position).getAudioDirPath());
 
     }
@@ -52,15 +70,15 @@ public class MainRecyclerViewAdapter extends RecyclerView.Adapter<MainRecyclerVi
     @Override
     public int getItemCount() {
         if(contents != null)
-        return contents.size();
+            return contents.size();
         return  0;
     }
 
     public static class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
-         ImageView imageView;
-         TextView textView;
-         String audioDirPath;
-         CardView cardView;
+        ImageView imageView;
+        TextView textView;
+        String audioDirPath;
+        CardView cardView;
 
         public void setAudioDirPath(String audioDirPath) {
             this.audioDirPath = audioDirPath;
