+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/database/ContentDAO.java
@@ -4,6 +4,7 @@ import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 import androidx.room.Update;
 import java.util.List;
@@ -17,9 +18,9 @@ public interface ContentDAO {
 
 
     @Query("DELETE FROM content_table")
-    public void deleteAll();
+    void deleteAll();
 
-    @Insert
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
     void insertContent(Content content);
 
     @Update
@@ -27,7 +28,6 @@ public interface ContentDAO {
 
     @Delete
     void deleteContent(Content content);
-
