+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/detection/Approach.java
@@ -1,5 +1,5 @@
 package com.pwc.commsgaze.detection;
 
 public enum Approach {
-    OPEN_CV_SPARSE_FLOW
+    OPENCV_SPARSE_FLOW
 }
