--- ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/detection/gazeutils/GazeStatus.java
@@ -1,6 +1,0 @@
-package com.pwc.commsgaze.detection.gazeutils;
-
-/*Describes the status of the gaze*/
-public enum GazeStatus {
-    ON_THE_WAY_TO_NEUTRAL,LEFT,RIGHT,UNKNOWN,NEUTRAL;
-}
