+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/detection/utils/GazeStatus.java
@@ -1,0 +1,6 @@
+package com.pwc.commsgaze.detection.utils;
+
+/*Describes the status of the gaze*/
+public enum GazeStatus {
+    ON_THE_WAY_TO_NEUTRAL,LEFT,RIGHT,UNKNOWN,NEUTRAL;
+}
