+++ ProjectSpace/commsgaze/src/test/java/com/pwc/commsgaze/ExampleUnitTest.java
@@ -2,7 +2,7 @@ package com.pwc.commsgaze;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
